detour: unexport serialization functions
